+++ composeApp/src/androidMain/kotlin/com/lastaosi/mycat/MyCatApplication.kt
@@ -1,9 +1,11 @@
 package com.lastaosi.mycat
 
 import android.app.Application
+import com.google.android.gms.ads.MobileAds
 import com.google.android.libraries.places.api.Places
 import com.lastaosi.mycat.di.appModule
 import com.lastaosi.mycat.di.databaseModule
+import com.lastaosi.mycat.util.NotificationHelper
 import com.lastaosi.mycat.worker.WorkManagerScheduler
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
@@ -24,9 +26,9 @@ class MyCatApplication : Application() {
                 appModule
             )
         }
-
+        NotificationHelper.createChannels(this)
         // WorkManager 스케줄 등록
         WorkManagerScheduler.scheduleVaccinationReminder(this)
-//        WorkManagerScheduler.scheduleMedicationReminder(this)
+        MobileAds.initialize(this)
     }
 }